ss1-backend-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

SQLAlchemy==2.0.44
asyncpg==0.30.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

boto3==1.34.120
python-dotenv==1.0.1

greenlet==3.3.0